Update colorGame.html
These changes ads the HTML part of the game to prepare it for the addition of a third difficulty level in the game, with 9 colored squares. The new difficulty level is called "Hard" and the previous "Hard" level is now "Medium". The game will start at the medium difficulty level.