colorGame.css - buttons and squares pointers
These changes are made to give pointer cursors to the top buttons (New Colors/Play Again, Easy, Medium and Hard). Also, a new class squarePointer will be toggled using javascript so the squares have a pointer cursor at the beginning of the game, and the cursor will go back to the original cursor as the squares as clicked and fade out, or when the correct square is clicked and the class is deactivated from all squares at once.